+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * [BREAKING] Rename string validator `present` -> `not_empty`. Rename error variant `Missing` -> `Empty`.
 * Introduce `new_unchecked` feature flag, that allows to bypass sanitization and validation.
 * Support derive of `JsonSchema` of `schemars` crate (requires `schemars08` feature).
+* Support string validation with `regex` (requires `regex1` feature).
 
 ### v0.1.1 - 2023-02-11
 * Initial release